% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParams.R
\name{readParams}
\alias{readParams}
\title{readParams}
\usage{
readParams(default_args, args = commandArgs(TRUE), print.input = TRUE)
}
\arguments{
\item{default_args}{data.frame with four columns: variable_name, default_value, description, preprocessing_function}

\item{args}{the result of 'commandArgs(TRUE)'}

\item{print.input}{either TRUE or FALSE to show the parsed input}
}
\description{
Read the parameters from command line (via commandArgs()) and return a list.
}
